dlt_pipeline/requirements.txt
dlt[bigquery]
google-cloud-storage
requests
